tests/wpt/web-platform-tests/2dcontext/shadows/2d.shadow.attributes.shadowBlur.valid.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>Canvas test: 2d.shadow.attributes.shadowBlur.valid</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/2dcontext/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/2dcontext/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.shadow.attributes.shadowBlur.valid</h1>
<p class="desc"></p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("");
_addTest(function(canvas, ctx) {

ctx.shadowBlur = 1;
_assertSame(ctx.shadowBlur, 1, "ctx.shadowBlur", "1");

ctx.shadowBlur = 0.5;
_assertSame(ctx.shadowBlur, 0.5, "ctx.shadowBlur", "0.5");

ctx.shadowBlur = 1e6;
_assertSame(ctx.shadowBlur, 1e6, "ctx.shadowBlur", "1e6");

ctx.shadowBlur = 0;
_assertSame(ctx.shadowBlur, 0, "ctx.shadowBlur", "0");


});
</script>
